Backend/requirements.txt
# Core Flask Framework
Flask==3.0.3
Flask-Cors==5.0.0
Flask-WTF==1.2.2

# JWT Authentication
Flask-JWT-Extended==4.6.0

# Database (PostgreSQL + SQLAlchemy)
SQLAlchemy==2.0.34
SQLAlchemy-Utils==0.41.0
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.1

# Supabase Client
supabase==2.10.0

# Form & Security Handling
WTForms==3.1.2
Werkzeug==3.0.3
